backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
openai
prophet
scikit-learn
cryptography
python-multipart
reportlab
groq
